backend/requirements.txt
# %%% Web Framework %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
fastapi==0.115.0
uvicorn==0.30.6
# %%% Data & ML %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
pyarrow==18.1.0
# %%% AI %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
openai==1.75.0
httpx==0.28.1
# %%% Auth %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
PyJWT==2.9.0
bcrypt==4.2.1
authlib==1.3.0
email-validator==2.2.0
# %%% Database %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
motor==3.7.1
pymongo==4.12.1
# %%% Validation & Serialization %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
pydantic==2.12.5
pydantic[email]==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.9
# %%% Utils %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
anyio==4.12.1
